Improved implementation and enabled 'test_client_limit_error-t' for CI
